axon-python-backend/axon-python-backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.11.0+cpu
fastapi
uvicorn
sentence-transformers
openai
pydantic
python-dotenv
